pyproject.toml
[project]
name = "event-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
]
